Move utils in DocumentDeclaration class
