+++ scripts/validation/validate.js
@@ -11,7 +11,6 @@ import * as services from '../../src/app/services/index.js';
 import fetch from '../../src/app/fetcher/index.js';
 import filter from '../../src/app/filter/index.js';
 import serviceSchema from './service.schema.js';
-import { extractCssSelectorsFromDocumentDeclaration } from '../../src/app/utils/index.js';
 
 const fs = fsApi.promises;
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -64,7 +63,7 @@ let servicesToValidate = args.filter(arg => !arg.startsWith('--'));
                 const document = await fetch({
                   url: location,
                   executeClientScripts,
-                  cssSelectors: extractCssSelectorsFromDocumentDeclaration(service.getDocumentDeclaration(type))
+                  cssSelectors: service.getDocumentDeclaration(type).getCssSelectors()
                 });
                 content = document.content;
                 mimeType = document.mimeType;
@@ -119,7 +118,7 @@ let servicesToValidate = args.filter(arg => !arg.startsWith('--'));
                   const document = await fetch({
                     url: location,
                     executeClientScripts,
-                    cssSelectors: extractCssSelectorsFromDocumentDeclaration(service.getDocumentDeclaration(type))
+                    cssSelectors: service.getDocumentDeclaration(type).getCssSelectors()
                   });
 
                   const secondFilteredContent = await filter({
