+++ src/app/services/documentDeclaration.test.js
@@ -1,17 +1,15 @@
 import chai from 'chai';
 
-import { extractCssSelectorsFromDocumentDeclaration } from './index.js';
+import DocumentDeclaration from './documentDeclaration.js';
 
 const { expect } = chai;
 
-describe('Utils', () => {
-  describe('#extractCssSelectorsFromDocumentDeclaration', () => {
+describe('DocumentDeclaration', () => {
+  describe('#extractCssSelectors', () => {
     context('With "select" property', () => {
       context('With string selector', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
-            contentSelectors: 'body'
-          });
+          const result = new DocumentDeclaration({ contentSelectors: 'body' }).getCssSelectors();
 
           expect(result).to.deep.equal([ 'body' ]);
         });
@@ -19,12 +17,12 @@ describe('Utils', () => {
 
       context('With range selector', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             contentSelectors: {
               startBefore: '#startBefore',
               endBefore: '#endBefore'
             }
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ '#startBefore', '#endBefore' ]);
         });
@@ -32,7 +30,7 @@ describe('Utils', () => {
 
       context('With an array of mixed selectors', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             contentSelectors: [
               {
                 startBefore: '#startBefore',
@@ -40,7 +38,7 @@ describe('Utils', () => {
               },
               'body'
             ]
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ '#startBefore', '#endBefore', 'body' ]);
         });
@@ -50,9 +48,9 @@ describe('Utils', () => {
     context('With "remove" property', () => {
       context('With string selector', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             noiseSelectors: 'body'
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ 'body' ]);
         });
@@ -60,12 +58,12 @@ describe('Utils', () => {
 
       context('With range selector', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             noiseSelectors: {
               startBefore: '#startBefore',
               endBefore: '#endBefore'
             }
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ '#startBefore', '#endBefore' ]);
         });
@@ -73,7 +71,7 @@ describe('Utils', () => {
 
       context('With an array of mixed selectors', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             noiseSelectors: [
               {
                 startBefore: '#startBefore',
@@ -81,7 +79,7 @@ describe('Utils', () => {
               },
               'body'
             ]
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ '#startBefore', '#endBefore', 'body' ]);
         });
@@ -91,10 +89,10 @@ describe('Utils', () => {
     context('With both "select" and "remove" property', () => {
       context('With string selector', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             contentSelectors: 'body',
             noiseSelectors: 'h1'
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ 'body', 'h1' ]);
         });
@@ -102,7 +100,7 @@ describe('Utils', () => {
 
       context('With range selector', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             contentSelectors: {
               startBefore: '#startBefore',
               endBefore: '#endBefore'
@@ -111,7 +109,7 @@ describe('Utils', () => {
               startBefore: '#startBefore',
               endBefore: '#endBefore'
             }
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ '#startBefore', '#endBefore', '#startBefore', '#endBefore' ]);
         });
@@ -119,7 +117,7 @@ describe('Utils', () => {
 
       context('With an array of mixed selectors', () => {
         it('extracts selectors', async () => {
-          const result = extractCssSelectorsFromDocumentDeclaration({
+          const result = new DocumentDeclaration({
             contentSelectors: [
               {
                 startBefore: '#startBefore',
@@ -134,7 +132,7 @@ describe('Utils', () => {
               },
               'body'
             ]
-          });
+          }).getCssSelectors();
 
           expect(result).to.deep.equal([ '#startBefore', '#endBefore', 'body', '#startBefore', '#endBefore', 'body' ]);
         });
